Added some protections to setup script

+++ setup.sh
@@ -4,6 +4,10 @@ for f in .[a-zA-Z]* ; do
   if [[ $f == ".git"* ]] ; then continue ; fi
   cd $HOME
   if [ -f $f ] ; then
+    if [ -f $f.orig ] ; then
+      echo "About to replate $f.orig, please remove to continue."
+      exit 1
+    fi
     mv $f $f.orig
   fi
   ln -s $OLDPWD/$f $f
@@ -11,7 +15,9 @@ for f in .[a-zA-Z]* ; do
 done
 cd $HOME
 cp $OLDPWD/.gitconfig .
-ln -s $OLDPWD/scripts scripts
+if [ ! -x $HOME/scripts ] ; then
+  ln -s $OLDPWD/scripts scripts
+fi
 if [ ! -d $HOME/src ] ; then mkdir $HOME/src ; fi
 for f in $OLDPWD/src/* ; do
   ln -s $f src
